Changed price data type from integer to double

+++ FinSight-Backend/src/main/java/com/FinSight_Backend/dto/PortfolioDTO.java
@@ -13,9 +13,9 @@ public class PortfolioDTO {
     private String name;
     private String description;
     private Boolean active;
-    private Long total_value;
-    private Long cost_basis;
-    private Integer yield;
+    private Double total_value;
+    private Double cost_basis;
+    private Double yield;
     private Integer user_id;
     private List<StockEntryDTO> stock_entries;
